Added core project without gfx backend or pipeline
I removed graphics backend from the project beetroot engine reinventing the wheel
as I want to try something slightly different for the renderer and allow myself to be more free with library usage. I still plan on writing code from scratch for that project but I would like to be faster at get things implemented for this variation of the engine.

I also removed pipeline (for now) I want to revisit this and remove the heavy texture compression lib I was using on the old project. the ISPC texture compressor can do all of the formats I'm interested in so I think I will end up using this and some lighter weight image loading lib as I don't need to support all the texture format. also I would like to start using KTX & GLTF over using my home grown DDS lib, I still plan on writing my own human readable & easily mergeable file format but this turned out a significantly bigger undertaking than I initially anticipated.

Refactored part of the windowing library to be based on function args instead of #defines
this does mean I need to convert chars to wchars which is a bit of a shame but it only happens once,
I've set a hard limit of 64 on the title name so I don't need to scan the array twice to get the real size of the converted char to wchar string.

Added submodule glm for maths lib

I considered making all of the libraries submodules of core or some new engine lib but I don't really see a point to that this time.