[GFX] Setup window resize
on `VK_ERROR_OUT_OF_DATE_KHR` we call window gfx_window_resize() which waits and the cleans up variosu resources and then rebuilds them. I was goinf to add an arg to all of the create & cleanup functions which would decide if something should be allocated/freed/invalidated but I think the better course of action would be to move all of the dynamic arrays to use pools with a max pool size and then just call vk alloc/free commands as in reality the CPU doesn't "own" these resources.