[GFX] Added Asset DB
Added asset DB for camera and transforms
from this I can now access the camera info and write a "script" to control the camera.

There are many other types that need to be moved to the DB i.e. the mesh data, descriptors material info etc etc.

I'm not 100% sold on the DB being the GFX system as many things that will use the DB don't need to drag in the entire GFX system but for now I think this is fine. I expect there will end up being a libDB which many of the systems will pick up and hook into.