[GFX] Added octahedral skybox
Added shader for rendering octahedral skyboxes.
While doing this I also updated the lit shader to use the uniform buffer instead of the push constants for view-proj and just a general bit of shader clean-up.

I would like to have a buffer of transform data stored on the gpu and on asset update we can just run a compute shader to update all of the assets that changed this frame.

Added the camera position to the ubo new renamed SceneUBO.

Added a new sampler, added a octahedral dds w/ mip maps in bc6h, this will be moved to a packaging step so we don't need to store converted assets on disk.

Added some debug build only texture info so I can inspect texture data via imgui. I may consider a parallel array for this in the future if this ends up getting much larger.