[CONVERTER] Added SPIRV to converter lib
I've wanted to move away from funning an external exe for shader compile for a while.

I've started the outline of a convert on demand library, I would like to hook this up to a dev/editor version of the engine which auto compile shader changes when a change is detected.

the code is a bit explicit about file loading. could do with some shared loader code with the engine but this is fine for now.

Added submodules
- glslang
-SPIRV-Headers
-SPIRV-Tools

only minor engine/editor changes will need to be put in place for on demand shader compile.