[CORE] Updated asset path to match cross project
Updated asset path, this will make a 1:1 raw -> converted check much easier as the paths with match between pipeline/runtime.
it should be as simple as adding the shader converter at the start of
the gfx_load_shader or gfx_load_shader_binary func. we will compile out this feature for non debug builds or potentially debug builds become the editor project and release becomes the runtime project

this is something I wanted to do a bit later on so the codebases wouldn't diverage too much.

will be the first major decision is the editor going to be built in the engine or will it just use the same base libs.

I'm leaning towards editor is in the engine but that could be considered a mistake.