[GFX] Setup the converter lib
Converter lib checks raw shaders and checks if they are more recent than what has already been converted, and if the raw is more recent we stall and compile shaders. I would like to get this setup that I can press a butt on a hotkey to try and reload/recompile and shaders and hotload them into the engine.