[GFX] Added line rendering
When doing GLTF parsing I need a way to check the orientation / origin of a object, so I've added a basic line rendering API.

Added pool for line rendering where the entity points to a span of points, and then we only send N points to the large GPU pool of points.

the current API only supports passing in 2 points but I've tested lines with N points and it works just fine.

Each line entity also stores a width to render that line with.

Added transform gizmo

next thing to do is get the smooth lines EXT working along with multi-point API