[GFX] Added N buffering for line rendering
Fixed up the code surrounding line rendering as it wasn't N buffering, because of this it was possible to write to the buffer while it was being read.

No longer write to a pool of points and then copy it over to the mapped memory every frame we instead just write directly to the mapped memory when we push a point.

I added a max count of 8 frames in flight, we still target frames in flight min + 1 which in most cases will be tripple buffered.

I also I ended up moving the swapChainImageCount out of the swapchain and into the root of the vulkan backend struct.

We pool alloc the max size of 8 but this is fine as vulkan just uses handles for these pools and we only allocate the memory when it's needed.

this also meant I could remove 2 more instances of mem_zalloc which is nice.

I still have a few instances of the STL being used in the codebase, I ran into vector usage in gfx_line which I would like to remove.