[IN_DEV] Investigating rolling my own GLTF parser
I was reading the spec and I thought that the GLTF standard looked simple enough that I could write a parser for what I need.

I want to be somewhat opinionated here i.e. I'm writing to an intermediate/in memory representation of the GLTF JSON, that way when it's time to do packaging it's easy to do all the index based indirection. it feels like this is the way GLTF should work in external libraries but they only provide a thin wrapper around JSON which is a shame.

I did pull in rapidJSON as it's nice and fast, it's not a lib I've used before but it seems easy enough to work with especially considering GLTF produces pretty flat JSON.

I have minimal implementation for.
asset
accessors
bufferViews
textures
samplers
images

Still to do.
buffers
meshes
materials
nodes
scenes
scene

unrelated to all this I also finally added
 add_definitions(-D_CRT_SECURE_NO_WARNINGS)
to the cmake, as lots of code complains about C functions. that don't use the _s or _n variants.