[IN_DEV] GLTF Support for interleaved buffer data + Sponza loaded.
Parse mesh data for Sponza \(^-^)/ YAY.

Many of the GLTF sample assets use buffers with interleaved data
this means I cannot just do a simple memcpy on a buffer region, I need to offset the cpy by the stride + the buffer byte offset  + accessor byte offset .

To get sponza working I added a few missing fields to the parse and fixed a bug where I was only writing 1 primitive, this meant I was constantly overwriting buffer at slot 0, this was fine on many of the test scenes as I only had 1 prim per mesh but it fell over when loading Sponza.

Added generic parse_gltf_texture func as we parse this json type pretty often, also added some new enums for alpha_mode, topology_mode min_filter & max_filter. These defaults (if any) are defined by the spec.

The next thing to do is work out what I want to do with the texture data & make each primitive exist in local space and apply an offset to the transform but this requires a bit more infra to exist first.

I want to compress it and save it off a .dds on disk so if it's already been converted I don't need to do this every time.

I'm in 2 minds about how I want the engine to consume all this data.
I really like being able to bundle up related data into a single "package" it means that everything needed for a "level" is totally self contained, but this does mean re-packaging all the data every time there is a change, it may be quite nice to have a more granular approach where instead of packaging up all the data I process it into a format the engine likes but all the files are "loose" on disk. so hot-reloading a mesh that's changed is as "simple as" unloading the existing mesh and loading in the updated one, this approach would require some level / DB file which is written out to disk and contains all the assets needed per level.