[GFX] Multisample resolve depth/stencil for debug line rending
Added depth to the line rendering util so if alpha is set to 0 for the line rendering it will use a default of full alpha if not behind geo and alpha of 30% when behind geo.

Hooked up the translate gizmo rending to the manipulator code path and took it out of the GUI db codepath as it was only really there out of convenience, expose a current selected type & pool index from the GUI db code.

there is a big HACK in this code in relation to the multisample resolve code, I already had the VK_KHR_depth_stencil_resolve ext enabled but this isn't playing nice with VK_KHR_dynamic_rendering it's throwing validation errors on depth / stencil resolve, with colour only validation failure, this would insinuate that the extension is not setup correctly.

when not using multisampling we go down a BLIT code path instead of image resolve which disables the gfx debug validation message hack as the blit works fine.

I would like to re-write much of the translate gizmo intersection tests into a standard API