[GFX] Immediate mode triangle strip rendering for Gizmos
Started building a triangle strip rendering API I'm planning to use it pretty much only for gizmos and other debug related work so I don't mind it spending a decent amount of time per frame uploading data to the GPU as this work won't be performed in release.

I'm using this API it to fill in the color of the rotation gizmo as I can't do that with a line as lines billboard to the camera.

I generally think it would be good to build up some set of immediate mode rendering APIs for debug too. i.e draw basic shapes where all you do is push a transform, colour & prim type and then the transform is applied to a pre-calculated list of triangles in some LUT/buffer that always exists on the GPU,
but for now I can build out this API using the triangle strip API even if it does potentially duplicate lots of geo.

There are lots of potential ways of optimising this but I will do that if and when it is ever needed.

TODO: remove std::vector from triangle strip add func as no where in the release code do we use std::vector (we do in GLTF parsing code but technically that is IN_DEV and will live in the pipeline code when it's done and only be included into the editor for on demand compile