[CORE] Building out more Ray Primitive tests
Ray plane & rect intersections and a debug draw grid, along with drawing a cube at the plane hit position.

Just adding these tests as and when I need them.

when I'm happy with the amount of tests I have I would like to refactor the API to either take in a Hit struct or return a Hit struct.
the current API feels awkward.