[CORE] Added dot test to gizmo & hit test ordering
There were a few issues with the translate gizmo mainly that I could interact with multiple handles at the same time if a ray overlapped multiple, so I've re-organised the code do that it is possible to do some calculations of the closest hit test.

Fixed a draw grid bug I introduced while tidying up code

added some util functions for mat4 so it's easy to extract info from a mat4 i.e. pos, rotation, scale.

the world space toggle is still IN_DEV it works correctly for the 2 axis handle, which is surprising I was expecting to just not apply rotations to the model matrix and just have it all work but I have a bug in the closest point on line logic I think.

I've also update the ray hit code to take in a Hit struct it just makes it less clunky to use