[CORE] Ray -> Plane -> Arc Intersection test for rotation gizmo
Added some code that computes the
plane -> !innerCircle -> outerCircle intersection
from this I can tell if the mouse is over one of the rotate Arcs.

I moved some code around so it could be used in the rotate function.

setup the rotation gizmo orientation code so it also works with a dot test against the camera.

I'm not happy with all the drawing APIs i.e. Arc & polyline. they currently require you to provide a model matrix which is a bit of a shame, I should just provide a up and normal vector and compute the orientation from that instead.