[CORE] Added scale gizmo + more transform functions.
Added transform scale local, world & another impl which takes a flag for world or local.

Added draw cube function, line points verts came from this post
https://stackoverflow.com/questions/28375338/cube-using-single-gl-triangle-strip

Fixed copy paste bug in transform_model_matrix_rotation this function was extracting the position instead of the rotation, I noticed this when I wrote the transform_translate_local function.

minor: Switched out the mat4f to braced initialisers
minor: Added vec3 include to transform.h
minor: Renamed widget_manipulate_transform to widget_manipulate_translate

I should consider refactoring the translate,rotate,scale to be stand alone functions which draw a gizmo based for any transform. I would need to remove the globals from the functions and the translate plane function.

in general there are quite a few useful function that are local to widget_manipulate which could be extracted. I would like to make a nicer to use API for translating a position on a plane, currently this requires a ray but it could be a case of translate on plane based on local direction in plane space.