[SHARED] platform target based feature system
I had a case where I had the define wrapped around some code but it was always staying defined out this was because the define itself wasn't included in the header, so now I can just wrap this in a CHECK_FEATURE and if FEATURE_ON or FEATURE_OFF isn't set for that define it will case a compile error.

this doesn't allow for nested conditions but I'm fine with this for now.