[GFX] Added engine consumable package from GLTF intermediate
Instead of parsing the GLTF and passing back just a mesh, I've setup the start of a asset package struct, for now it only contains GFX ready Mesh data and texture path info. but this will expand to support more.

once I'm happy with the format & functionality I will rewrite all this code in an easy to serialise format. i.e.  a header of offsets & counts and C arrays of data