[GFX] Fixed default cube normals & shader normal
I never setup the normal for the cube or getting the world space normal, I also wrote a quick hacky script for displaying the position of the hardcoded light positions from, once there is a way to upload light positions this code will update to display those positions.

there is also currently no lighting entity, which is a bit annoying,
I would like a way to select a light and move it around.