[GFX] Fixed immediate mode alpha test z fighting & added more immediate primitives
Added box line & box poly
added frustum poly which is really just a variant on the box.

also the immediate mode poly rendering was using the line shader, so I fixed that and added some very basic code to do a dither. I would like to include a param on the poly to tell the renderer if it should use the dither shader or not