[GFX] Refactored db pool
the db / pool allocation was getting a bit unruly.

I opted to use templates here as it makes the code easy to use & add new pools, I could have done this in C but it would have been a tad more awkward to use as I would need to create a new pool struct for
 every new pool ( which is what I was doing before this change but just inlined into the static alloc ).

I've ensured that there are no templates leaking out and there is type erasure for the pool info which is used to display the allocation table.

moved away from static allocation on boot to an explicit create function call.

may change all the pool sizes from defines to constexpr but it's fine for now.