[CORE] Removed unused engine system
this callback based engine system has never been used in the current project, it is a hold over from a old version of the engine.

if I need this again I will probably just make it an array instead of a map, a map seems a bit overkill for iterating over systems by index.